requirements.txt
python-dotenv
langgraph
langchain-openai
openai
matplotlib
langgraph-checkpoint-postgres
psycopg[binary]
snowflake-connector-python
faiss-cpu